mainly ui and darkmode changes/ with js cleanup
