+++ src/css/darkmode.css
@@ -1,0 +1,57 @@
+.darkmode {
+  background-color: #0b0505;
+  color: #fff;
+}
+.darkmode .dark-mode-body {
+  background-color: #0b0505;
+}
+.darkmode .fact {
+  background-color: #808080;
+}
+.darkmode button {
+  background-color: #ff981f;
+  color: #0b0505;
+  font-weight: 600;
+  letter-spacing: 1.1px;
+}
+.darkmode button:active {
+  box-shadow: 0px 0px 5px 2px #fff;
+  transform: scale(1.2);
+}
+.darkmode button:focus,
+.darkmode button:hover {
+  cursor: pointer;
+  transform: scale(0.95);
+}
+.darkmode .dark-inputs {
+  background-color: transparent;
+}
+.darkmode .dark-dropdown {
+  background-color: #6b6b6b;
+}
+.darkmode .search-bar-container {
+  background-color: rgba(41, 36, 36, 0.7843137255);
+  color: black;
+  font-size: 2rem;
+}
+.darkmode ::-moz-placeholder {
+  color: #fff;
+}
+.darkmode :-ms-input-placeholder {
+  color: #fff;
+}
+.darkmode ::placeholder {
+  color: #fff;
+}
+.darkmode footer {
+  background-color: #0b0505;
+}
+.darkmode select {
+  background-color: rgba(41, 36, 36, 0.7843137255);
+  color: #fff;
+}
+.darkmode option,
+.darkmode #placeholder-select {
+  background-color: #6b6b6b;
+  color: #fff;
+}/*# sourceMappingURL=darkmode.css.map */
